refactor(app): migrate restore/select and admin info orchestration
